+++ packages/components/src/components/heading/heading-styles.ts
@@ -8,7 +8,6 @@ import {
   fontSizeHeadingSmall,
   fontSizeHeadingXLarge,
   fontSizeHeadingXXLarge,
-  fontSizeHeadingXXXLarge,
   headingXXLargeStyle,
 } from '@porsche-design-system/utilities-v2';
 import { getTypographySlottedJssStyle, getTypographyRootJssStyle } from '../../styles/typography-styles';
@@ -19,7 +18,6 @@ export const sizeMap: { [key in Exclude<HeadingSize, 'inherit'>]: string } = {
   large: fontSizeHeadingLarge,
   'x-large': fontSizeHeadingXLarge,
   'xx-large': fontSizeHeadingXXLarge,
-  'xxx-large': fontSizeHeadingXXXLarge,
 };
 
 export const getComponentCss = (