+++ packages/utilities/projects/utilities/src/js/font/fontSizeHeading.ts
@@ -3,7 +3,6 @@ import { fontSizeHeadingMedium } from './fontSizeHeadingMedium';
 import { fontSizeHeadingLarge } from './fontSizeHeadingLarge';
 import { fontSizeHeadingXLarge } from './fontSizeHeadingXLarge';
 import { fontSizeHeadingXXLarge } from './fontSizeHeadingXXLarge';
-import { fontSizeHeadingXXXLarge } from './fontSizeHeadingXXXLarge';
 
 export const fontSizeHeading = {
   small: fontSizeHeadingSmall,
@@ -11,5 +10,4 @@ export const fontSizeHeading = {
   large: fontSizeHeadingLarge,
   xLarge: fontSizeHeadingXLarge,
   xxLarge: fontSizeHeadingXXLarge,
-  xxxLarge: fontSizeHeadingXXXLarge,
 };