+++ packages/utilities/projects/utilities/src/js/typography/heading/headingXXXLargeStyle.ts
@@ -1,6 +1,0 @@
-import { _headingFontPartA, _headingFontPartB } from './headingShared';
-import { fontSizeHeadingXXXLarge } from '../../font';
-
-export const headingXXXLargeStyle = {
-  font: `${_headingFontPartA}${fontSizeHeadingXXXLarge}${_headingFontPartB}`,
-} as const;